src/project/requirements.txt
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.53.1
iconify==0.0.103
idna==2.10
imutils==0.5.4
joblib==1.4.2
kids.cache==0.0.7
kiwisolver==1.4.5
libsass==0.23.0
lxml==5.3.0
matplotlib==3.9.0
mock==5.1.0
mysql-connector-python==9.0.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
perlin-noise==1.13
pillow==10.4.0
pycparser==2.22
pyinstall==0.1.4
PyMySQL==1.1.1
pyparsing==3.1.2
PyQt5==5.14.2
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
pyserial==3.5
PySide2==5.15.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
QT-PyQt-PySide-Custom-Widgets==0.3.8
QtPy==2.4.1
qtsass==0.4.0
requests==2.23.0
scikit-learn==1.5.1
scipy==1.14.0
shiboken2==5.15.2.1
shiboken6==6.7.2
six==1.16.0
termcolor==2.4.0
threadpoolctl==3.5.0
tinycss2==1.3.0
typing==3.7.4
tzdata==2024.1
urllib3==1.25.11
watchdog==4.0.2
webencodings==0.5.1
